refactor(schema): move saveChangeset to schema root and remove re-export shims
Move saveChangeset from push/changeset to the shared schema/changeset module.
Delete the push/changeset and push/serialize re-export files and update all
consumers to import directly from the shared parent directory.
