--- packages/cli/src/commands/schema/push/serialize.ts
@@ -1,1 +1,0 @@
-export { serializeComponent, serializeComponentFolder, serializeDatasource } from '../serialize';
